+++ TetrisConsoleApplication/Tetris.Utilities/Constants.cs
@@ -36,7 +36,7 @@ namespace Tetris.Utilities
         public const string Space = "Space - Rotate tetrimino";
 
         public const int StartLevel = 1;
-        public const int StartScore = 0;
+        public const long StartScore = 0;
         public const int StartLinesCleared = 0;
 
         public const int TetrominoRefillCount = 10;
